fan_controller_raspbian/requirements.txt
RPi.GPIO
w1thermsensor
